requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx==0.27.2